:sparkles: Add Komponent spesific modules to page template
