+++ aksel.nav.no/website/app/_ui/portable-text/CustomPortableText.components.tsx
@@ -19,7 +19,8 @@ function customPortableTextComponents({
   const block = blockComponents({ typoConfig });
   return {
     block,
-    unknownBlockStyle: block.normal,
+    unknownBlockStyle: ({ children }) =>
+      withSanitizedBlock(<BodyShort spacing>{children}</BodyShort>),
   };
 }
