+++ aksel.nav.no/website/app/_ui/typography/Typography.module.css
@@ -7,3 +7,21 @@
   /* stylelint-disable-next-line csstools/value-no-unknown-custom-properties */
   font-family: var(--website-font-mono);
 }
+
+/* className="list-margin max-w-text" */
+.typoList {
+  max-width: 600px;
+
+  /* stylelint-disable-next-line aksel/no-class-override */
+  .navds-body-long.navds-typo--spacing + & {
+    margin-top: -1rem;
+  }
+}
+
+.typoListIcon {
+  transform: rotate(45deg);
+
+  & > * {
+    fill: var(--ax-bg-brand-blue-strong);
+  }
+}
